money-machine/src-python/requirements.txt
# Money Machine - Python Trading Engine

ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
aiohttp>=3.9.0
google-generativeai>=0.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
